pyproject.toml
[tool.poetry]
name = "content-gpt-backend"
version = "0.1.0"
description = ""
authors = ["Ammar <syed.ammar.98sep@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.6.2"
fastapi = "^0.110.0"
gunicorn = "^21.2.0"
fastapi-utils = "^0.2.1"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
uvicorn = "^0.29.0"
colorlog = "^6.8.2"
celery = "^5.3.6"
redis = "^5.0.3"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

